fix: `firstRun` should not be encapsulated
Otherwise, the functionality to restart the game does not work.
Because in the `start()`, we are initializing the game object and
previously in the game object, `firstRun` was by default set to true.
This causes problems as this is fine in the first run but on later
runs this prediction is incorrect.